group mqtt updates together and do a single sleep delay
Let the initial config publishing happen immediately,
but batch together the availability and other status
updates so that we can do a single sleep before we process
all of them.

This will make it easier to modularize the registration
of multiple entities with hass without littering the code
with sleeps.