adjust topics so that multiple hubs can be used on the same network
While we don't (currently?) support serving multiple hubs from the
same instance of pview, this change will allow multiple instances
to be running on the same network without collision in the topics
used to receive commands from hass.

Similarly, adjust the postback url so that the hub sends its serial
number as part of the URL that we register into it.

Add the ability to delete entities from hass; we have to delete
the old configuration before we can register the new configuration
with adjusted topics.