add extractors to mqtt handlers
Yes, this is a bit over-engineered just for this project, but
I'm planning on moving mqtt_helper into the mosquitto crate
more directly.

The extractor stuff is patterned after the similar concept in
the axum framework, where simply declaring parameters in your
handler, elements from the incoming request can be parsed out
and you only really pay for the bits you're using.